Remove redundant preference composables
