+++ app/src/main/java/eu/kanade/presentation/components/Divider.kt
@@ -1,0 +1,17 @@
+package eu.kanade.presentation.components
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+const val DIVIDER_ALPHA = 0.2f
+
+@Composable
+fun Divider(
+    modifier: Modifier = Modifier,
+) {
+    androidx.compose.material3.Divider(
+        modifier = modifier,
+        color = MaterialTheme.colorScheme.onSurface.copy(alpha = DIVIDER_ALPHA),
+    )
+}
