+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/extension/ExtensionFilterPresenter.kt
@@ -13,7 +13,6 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.flow.stateIn
 import logcat.LogPriority
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
@@ -36,7 +35,6 @@ class ExtensionFilterPresenter(
                     logcat(LogPriority.ERROR, exception)
                     _events.send(Event.FailedFetchingLanguages)
                 }
-                .stateIn(presenterScope)
                 .collectLatest(::collectLatestSourceLangMap)
         }
     }
