+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/source/SourcesFilterPresenter.kt
@@ -15,7 +15,6 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.flow.stateIn
 import logcat.LogPriority
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
@@ -40,7 +39,6 @@ class SourcesFilterPresenter(
                     logcat(LogPriority.ERROR, exception)
                     _events.send(Event.FailedFetchingLanguages)
                 }
-                .stateIn(presenterScope)
                 .collectLatest(::collectLatestSourceLangMap)
         }
     }
